Add getEventBySlug action and update page components
- Added getEventBySlug action to fetch event by slug
- Updated page components to use getEventsWithImageData and getEventBySlug actions
